detector/requirements.txt
pika
pymseed
numpy
scipy
obspy
matplotlib
psycopg2-binary
